moved data_handling.py, subsampling.py, and plotting.py into activity folder and renamed submodules
